+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Improvements:
 Bug fixes:
 * [OLMIS-6330](https://openlmis.atlassian.net/browse/OLMIS-6330): fixed trigerring reference-ui build.
 * [OLMIS-6357](https://openlmis.atlassian.net/browse/OLMIS-6357): Fixed logging into the application on Firefox with cleared history.
+* [OLMIS-6438](https://openlmis.atlassian.net/browse/OLMIS-6438): Fixed searching for orderables on products modal and in API console.
 
 5.5.1 / 2019-05-27
 ==================